[layout] update GiftPage html to components
